src/requirements.txt
celery==5.4.0
dj_database_url==2.2.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf_yasg==1.21.8
python-dotenv==1.2.1
redis==5.1.1
django-cors-headers==3.13.0
gunicorn>=21.0.0
whitenoise==6.6.0
psycopg2-binary==2.9.10
geocoder==1.38.1
